+++ src/lib/components/buttons/file-chooser/terra-file-chooser.component.ts
@@ -1,11 +1,10 @@
-import { Component, EventEmitter, Input, Output, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Input, Output, TemplateRef, ViewChild } from '@angular/core';
 import { L10nTranslationService } from 'angular-l10n';
-import { isNullOrUndefined } from 'util';
 import { TerraFileBrowserComponent } from '../../file-browser/terra-file-browser.component';
 import { TerraButtonComponent } from '../button/terra-button.component';
 import { TerraBaseStorageService } from '../../file-browser/terra-base-storage.interface';
 import { TerraStorageObject } from '../../file-browser/model/terra-storage-object';
-import { TerraOverlayComponent } from '../../layouts/overlay/terra-overlay.component';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { TerraOverlayButtonInterface } from '../../layouts/overlay/data/terra-overlay-button.interface';
 
 /**
@@ -25,11 +24,7 @@ export class TerraFileChooserComponent extends TerraButtonComponent {
     }
 
     public get inputPrimaryBrowserButtonCaption(): string {
-        if (!isNullOrUndefined(this._primaryBrowserButtonCaption) && this._primaryBrowserButtonCaption.length > 0) {
-            return this._primaryBrowserButtonCaption;
-        }
-
-        return this._translation.translate(this._translationPrefix + '.choose');
+        return this._primaryBrowserButtonCaption || this._translation.translate(this._translationPrefix + '.choose');
     }
 
     @Input()
@@ -38,11 +33,7 @@ export class TerraFileChooserComponent extends TerraButtonComponent {
     }
 
     public get inputSecondaryBrowserButtonCaption(): string {
-        if (!isNullOrUndefined(this._secondaryBrowserButtonCaption) && this._secondaryBrowserButtonCaption.length > 0) {
-            return this._secondaryBrowserButtonCaption;
-        }
-
-        return this._translation.translate(this._translationPrefix + '.cancel');
+        return this._secondaryBrowserButtonCaption || this._translation.translate(this._translationPrefix + '.cancel');
     }
 
     @Input()
@@ -52,13 +43,7 @@ export class TerraFileChooserComponent extends TerraButtonComponent {
     public inputAllowFolders: boolean = true;
 
     @Input()
-    public set inputStorageServices(services: Array<TerraBaseStorageService>) {
-        this._storageServices = services;
-    }
-
-    public get inputStorageServices(): Array<TerraBaseStorageService> {
-        return this._storageServices;
-    }
+    public inputStorageServices: Array<TerraBaseStorageService>;
 
     @Output()
     public outputSelected: EventEmitter<TerraStorageObject> = new EventEmitter<TerraStorageObject>();
@@ -76,27 +61,27 @@ export class TerraFileChooserComponent extends TerraButtonComponent {
         TerraFileBrowserComponent
     >();
 
-    @ViewChild('overlay', { static: true })
-    public overlay: TerraOverlayComponent;
-
-    @ViewChild('fileBrowser', { static: true })
+    @ViewChild('fileBrowser', { static: false })
     public fileBrowser: TerraFileBrowserComponent;
 
+    @ViewChild(TemplateRef, { static: true })
+    public _dialogTemplateRef: TemplateRef<any>;
+
+    /** @deprecated since v12. */
     public primaryOverlayButton: TerraOverlayButtonInterface;
 
+    /** @deprecated since v12. */
     public secondaryOverlayButton: TerraOverlayButtonInterface;
 
-    private _translationPrefix: string = 'terraFileInput';
+    public _translationPrefix: string = 'terraFileInput';
+
+    public _selectedObject: TerraStorageObject;
 
     private _primaryBrowserButtonCaption: string = '';
 
     private _secondaryBrowserButtonCaption: string = '';
 
-    private _selectedObject: TerraStorageObject;
-
-    private _storageServices: Array<TerraBaseStorageService>;
-
-    constructor(private _translation: L10nTranslationService) {
+    constructor(private _translation: L10nTranslationService, private _dialog: MatDialog) {
         super();
 
         this.primaryOverlayButton = {
@@ -105,7 +90,6 @@ export class TerraFileChooserComponent extends TerraButtonComponent {
             isDisabled: true,
             clickFunction: (): void => {
                 this.outputSelected.emit(this._selectedObject);
-                this.overlay.hideOverlay();
             }
         };
 
@@ -115,23 +99,31 @@ export class TerraFileChooserComponent extends TerraButtonComponent {
             isDisabled: false,
             clickFunction: (): void => {
                 this.outputCancelled.emit();
-                this.overlay.hideOverlay();
             }
         };
     }
 
     public onClick(event: Event): void {
         this.outputClicked.emit(event);
-        this.overlay.showOverlay();
+        const dialogRef: MatDialogRef<any> = this._dialog.open(this._dialogTemplateRef);
+        dialogRef.afterOpened().subscribe(() => {
+            this.onBrowserShow();
+        });
+        dialogRef.afterClosed().subscribe(() => {
+            this.onBrowserHide();
+        });
     }
 
     public onSelectedObjectChange(selectedObject: TerraStorageObject): void {
-        if (isNullOrUndefined(selectedObject) || selectedObject.isDirectory) {
-            this.primaryOverlayButton.isDisabled = true;
-        } else {
-            this.primaryOverlayButton.isDisabled = false;
-            this._selectedObject = selectedObject;
-        }
+        // workaround since change detection is not finished when selectedObject is set
+        setTimeout(() => {
+            if (!selectedObject || selectedObject.isDirectory) {
+                this.primaryOverlayButton.isDisabled = true;
+            } else {
+                this.primaryOverlayButton.isDisabled = false;
+                this._selectedObject = selectedObject;
+            }
+        });
     }
 
     public onBrowserShow(): void {