+++ src/lib/components/file-browser/image-preview/image-preview.component.ts
@@ -2,9 +2,8 @@ import { ChangeDetectorRef, Component, Inject, Input } from '@angular/core';
 import { TerraStorageObject } from '../model/terra-storage-object';
 import { TerraBaseStorageService } from '../terra-base-storage.interface';
 import { TerraImageMetadata } from '../model/terra-image-metadata.interface';
-import { isNullOrUndefined } from 'util';
 import { TerraBaseMetadataStorageService } from '../terra-base-metadata-storage.interface';
-import { L10nLocale, L10nTranslationService, L10N_LOCALE } from 'angular-l10n';
+import { L10N_LOCALE, L10nLocale, L10nTranslationService } from 'angular-l10n';
 
 @Component({
     selector: 'terra-image-preview',
@@ -26,7 +25,7 @@ export class TerraImagePreviewComponent {
         this._metadata = {};
         this._isLoading = true;
         if (
-            !isNullOrUndefined(object) &&
+            object != null &&
             this.inputStorageService &&
             this.inputStorageService instanceof TerraBaseMetadataStorageService
         ) {