+++ src/lib/terra-components.module.ts
@@ -3,7 +3,7 @@ import { components, exportedComponents } from './components/component-collectio
 import { directives, exportedDirectives } from './components/directive-collection';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { L10nTranslationModule } from 'angular-l10n';
+import { L10nIntlModule, L10nTranslationModule } from 'angular-l10n';
 import { MyDatePickerModule } from 'mydatepicker';
 import { RouterModule } from '@angular/router';
 import { QuillModule } from 'ngx-quill';
@@ -24,6 +24,10 @@ import { MatDatepickerModule } from '@angular/material/datepicker';
 import { A11yModule } from '@angular/cdk/a11y';
 import { MatInputModule } from '@angular/material/input';
 import { MatPseudoCheckboxModule } from '@angular/material/core';
+import { MatTreeModule } from '@angular/material/tree';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatTableModule } from '@angular/material/table';
+import { MatExpansionModule } from '@angular/material/expansion';
 import { TerraTooltipModule } from './components/tooltip/terra-tooltip.module';
 
 @NgModule({
@@ -51,6 +55,11 @@ import { TerraTooltipModule } from './components/tooltip/terra-tooltip.module';
         A11yModule,
         MatSelectModule,
         MatPseudoCheckboxModule,
+        MatTreeModule,
+        MatToolbarModule,
+        MatTableModule,
+        MatExpansionModule,
+        L10nIntlModule,
         TerraTooltipModule
     ],
     declarations: [...components, ...directives, ...pipes],