+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 12.4.0 (X.X.2021) - UNRELEASED
+
+### Features
+
+-   **terra-file-list** use components from `@angular/material` instead of our own deprecated ones.
+-   **terra-file-input** use components from `@angular/material` instead of our own deprecated ones.
+-   **terra-file-browser** use components from `@angular/material` instead of our own deprecated ones.
+-   **terra-image-preview** use components from `@angular/material` instead of our own deprecated ones.
+-   **terra-file-chooser** use dialog from `@angular/material` instead of our deprecated `terra-overlay`.
+
 # 12.3.0 (04.11.2021)
 
 ### Minor changes